Move 'z-index' and 'bullet' to the 'game' directory.
